# specs/002-frontend-saas-service/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Frontend SaaS Service API
  version: 1.0.0
paths:
  /users:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
  /users/{userId}:
    get:
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /knowledge_vaults:
    post:
      summary: Create a new knowledge vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeVault'
      responses:
        '201':
          description: Knowledge vault created successfully
  /blog_posts:
    post:
      summary: Create a new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPost'
      responses:
        '201':
          description: Blog post created successfully
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
    KnowledgeVault:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
    BlogPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vault_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        status:
          type: string